+++ plugins/rplanners/ParabolicPathSmooth/Math.h
@@ -12,8 +12,18 @@ typedef std::vector<Real> Vector;
 #define PARABOLICWARN RAVELOG_DEBUG
 #define PARABOLIC_ASSERT BOOST_ASSERT
 
+//tolerance for time
+const static Real EpsilonT = 1e-8;
+//tolerance for position
+const static Real EpsilonX = 1e-7;
+//tolerance for velocity
+const static Real EpsilonV = 1e-7;
+//tolerance for acceleration
+const static Real EpsilonA = 1e-6;
+
 //can replace this with your favorite representation/tests of infinity
 const static Real Inf = 1e300;
+
 inline bool IsInf(Real x) {
     return x==Inf;
 }